+++ addon/components/citation-widget/component.js
@@ -1,4 +1,9 @@
 import Ember from 'ember';
+
+import { task, timeout } from 'ember-concurrency';
+import config from 'ember-get-config';
+
+import { authenticatedAJAX } from 'ember-osf/utils/ajax-helpers';
 import layout from './template';
 
 /**
@@ -28,12 +33,32 @@ const citationStyles = [
  * @param {node} node for which to fetch citations
  */
 export default Ember.Component.extend({
+    store: Ember.inject.service(),
+    i18n: Ember.inject.service(),
+
     layout,
     apa: null,
     chicago: null,
     mla: null,
     node: null,
-    store: Ember.inject.service(),
+
+    selectedStyle: '',
+    citationText: '',
+
+    styles: Ember.A([]),
+
+    placeholderMessage: Ember.computed(function() {
+        return this.get('i18n').t('eosf.components.citationWidget.placeholderMessage');
+    }),
+    loadingMessage: Ember.computed(function() {
+        return this.get('i18n').t('eosf.components.citationWidget.loadingMessage');
+    }),
+    noMatchesMessage: Ember.computed(function() {
+        return this.get('i18n').t('eosf.components.citationWidget.noMatchesMessage');
+    }),
+    searchMessage: Ember.computed(function() {
+        return this.get('i18n').t('eosf.components.citationWidget.searchMessage');
+    }),
 
     didReceiveAttrs() {
         const node = this.get('node');
@@ -43,6 +68,7 @@ export default Ember.Component.extend({
         }
 
         const citationLink = node.get('links.relationships.citation.links.related.href');
+        this.set('citationLink', citationLink);
 
         for (const { linkSuffix, attr } of citationStyles) {
             this.get('store')
@@ -50,5 +76,39 @@ export default Ember.Component.extend({
                 .ajax(`${citationLink}${linkSuffix}/`, 'GET')
                 .then(resp => this.set(attr, resp.data.attributes.citation));
         }
-    }
+    },
+
+    actions: {
+        selectStyle(style) {
+            this.set('selectedStyle', style);
+            this.get('_selectStyle').perform(style.id);
+        }
+    },
+
+    _selectStyle: task(function* (id) {
+        const citationLink = this.get('citationLink');
+        const response = yield authenticatedAJAX({ url: `${citationLink}${id}/` });
+        this.set('citationText', response.data.attributes.citation);
+    }).restartable(),
+
+    findStyles: task(function* (term) {
+        yield timeout(500);
+        const response = yield Ember.$.ajax({
+            method: 'GET',
+            url: `${config.OSF.apiUrl}/${config.OSF.apiNamespace}/citations/styles/?filter[title,short_title]=${term}&page[size]=100`,
+            dataType: 'json',
+            contentType: 'application/json'
+        });
+        if (response.links.next !== null) {
+            response.data.push({
+                attributes: {
+                    // TODO: Can (ask product) lazy load the rest of the styles when scrolled down, once @adlius's PR is merged:
+                    // https://github.com/CenterForOpenScience/ember-osf/pull/338
+                    title: `${response.links.meta.total - 100} more, type more to narrow results`,
+                },
+                disabled: true,
+            });
+        }
+        return response.data;
+    }).restartable()
 });