+++ addon/utils/load-relationship.js
@@ -9,11 +9,36 @@ export function loadRelation(model, relationship) {
         model = model.get('content')
     }
 
+    // If model is null return a promise that never resolves
+    // and wait for the next call were model is resolved
     if (!model) return new Ember.RSVP.Promise(() => null);
 
     let results = Ember.A();
     let promise = loadAll(model, relationship, results).then(() => results);
     return ArrayPromiseProxy.create({promise});
+}
+
+export function loadPage(model, relationship, pageSize, page, options = {}) {
+    let query = {
+        'page[size]': pageSize || 10,
+        page: page || 1
+    };
+    query = Ember.merge(query, options || {});
+    Ember.set(model, 'query-params', query);
+
+    return model.queryHasMany(relationship, query).then(results => {
+        var remaining = 0;
+        if (results.meta) {
+            var total = results.meta.total;
+            var pageSize = results.meta.per_page;
+            remaining = total - (page * pageSize);
+        }
+        return {
+            results: results.toArray(),
+            hasRemaining: remaining > 0,
+            remaining: remaining,
+        };
+    });
 }
 
 export default function loadAll(model, relationship, dest, options = {}) {